Move more code from cmd2.py into utils.py
